+++ TransactionStore.Contracts/ITransactionRepository.cs
@@ -7,7 +7,7 @@ public interface ITransactionRepository
 {
     public Task<int> CreateTransactionAsync(TransactionEntity transaction);
     public Task<List<int>> CreateTransferTransactionAsync(TransactionEntity transferWithdraw, TransactionEntity transferDeposit);
-    public Task<int> GetAccountBalanceAsync(int accountId);
+    public Task<decimal> GetAccountBalanceAsync(int accountId);
     public Task<TransactionEntity> GetTransactionByIdAsync(int transactionId);
     public Task<List<TransactionEntity>> GetAllTransactionsByAccountIdAsync(int accountId);
 }
